+++ extras/tests/JsonObject/containsKey.cpp
@@ -29,7 +29,7 @@ TEST_CASE("JsonObject::containsKey()") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("key is a VLA") {
-    int i = 16;
+    size_t i = 16;
     char vla[i];
     strcpy(vla, "hello");