+++ extras/tests/JsonObject/createNestedArray.cpp
@@ -16,7 +16,7 @@ TEST_CASE("JsonObject::createNestedArray()") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("key is a VLA") {
-    int i = 16;
+    size_t i = 16;
     char vla[i];
     strcpy(vla, "hello");